♻️ refactor : 메인 하우잇워크 세션 분리   #288
